requirements.txt
Django==5.1.0
djangorestframework==3.15.2
django-channels==4.1.0
channels-redis==4.2.0
psycopg2-binary==2.9.9
dj-database-url==2.2.0
python-decouple==3.8
gunicorn==22.0.0
whitenoise==6.7.0